w_requirements.txt
aiohttp==3.10.2
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.1.0
#autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
boto3==1.34.69
botocore==1.34.69
celery==5.3.4
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dataclasses-json==0.6.4
decorator==4.4.2
distro==1.9.0
Django==4.2.15
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-crispy-forms==1.14.0
django-cron==0.6.0
django-decouple==2.1
django-filter==23.3
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
dynaconf==3.2.5
frozenlist==1.4.1
greenlet==3.0.3
gunicorn
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hvac==2.1.0
hyperlink==21.0.0
idna==3.7
imageio==2.31.6
imageio-ffmpeg==0.4.9
incremental==24.7.0
inflection==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
knox==0.1.14
kombu==5.3.2
langchain==0.2.15
langchain-community==0.2.9
langchain-core==0.2.35
langchain-text-splitters==0.2.0
langsmith==0.1.75
loguru==0.7.2
MarkupSafe==2.1.5
marshmallow==3.21.1
moviepy==1.0.3
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.0
numpy
openai==1.31.0
orjson==3.9.15
packaging==23.2
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
#psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.0
pydantic_core==2.16.3
PyJWT==2.4.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
requests==2.32.2
s3transfer==0.10.1
service-identity==23.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.31
sqlparse==0.5.0
tenacity==8.2.3
tqdm==4.66.3
Twisted==24.7.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3>=1.25.4,<2.3
validators==0.23.2
vine==5.0.0
wcwidth==0.2.9
win32-setctime==1.1.0
whitenoise
yarl==1.9.4
zope.interface==6.1
